+++ test/mocks/pull-request.mock.ts
@@ -1,0 +1,1 @@
+export default jest.fn().mockName("prs.ts::main");