+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Run Action on push events. ([#236])
 
 ## [1.0.2] - 2020-08-21
 
@@ -147,3 +147,4 @@ Versioning].
 [#207]: https://github.com/ericcornelissen/svgo-action/pull/207
 [#223]: https://github.com/ericcornelissen/svgo-action/pull/223
 [#239]: https://github.com/ericcornelissen/svgo-action/pull/239
+[#236]: https://github.com/ericcornelissen/svgo-action/pull/236